Update queryset after dispatching tasks
Before this patch, if you wanted to redispatch tasks with
`status=error`, the `queryset.update` would change their status, and the
`iterator()` query would receive no objects.

Now, we start a transaction, then call `queue_task`, which adds tasks to
celery _when the transaction commits_, then we update the queryset, and
when the transaction is done, the jobs are sent to celery.