Extract path parts before indexing
Currently, using the path_hierarcy tokenizer means we can't aggregate
on the field. That means we would have to set `"fielddata": true`, which comes
with a memory cost. This was discussed but not solved in elastic/elasticsearch#18064.

Perhaps fielddata would be OK for file paths like ours, but this seems safer.